Refactor simulated genesis startup sequences
