+++ ethereum/datastructures/src/main/java/tech/pegasys/artemis/datastructures/util/MockStartBeaconStateGenerator.java
@@ -45,6 +45,7 @@ public class MockStartBeaconStateGenerator {
     }
     final BeaconStateWithCache initialState =
         BeaconStateUtil.initialize_beacon_state_from_eth1(BLOCK_HASH, genesisTime, deposits);
+    initialState.setGenesis_time(genesisTime);
     return initialState;
   }
 }
